简化选中路径添加按钮：直接选目录并移除弹出菜单。
